Fix in Unicode.cpp, fixed some warnings
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@722 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/client/CChatBox.cpp
@@ -172,7 +172,7 @@ lines_iterator CChatBox::At(int i)  {
 	if (i <= 0)
 		return WrappedLines.begin();
 
-	if (i >= WrappedLines.size())
+	if (i >= (int)WrappedLines.size())
 		return WrappedLines.end();
 
 	// Go to the right iterator
